app/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
#opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
Pillow==8.0.1
pandas==1.2.1
streamlit==1.10
streamlit-webrtc==0.42
matplotlib==3.5.1
tqdm==4.56.0
PyYAML==5.3.1
seaborn==0.11.1
protobuf==3.20.1
click==7.1.2
pyOpenSSL
aiortc==1.3.2
numpy==1.22.3
pydub==0.25.1
typing_extensions==4.1.1
pytz